feat: DSButton으로 네이밍 수정 후 small 버튼 타입 추가
